% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compare034_1sample_sigma.R
\name{boxplot_1sample_sigma}
\alias{boxplot_1sample_sigma}
\title{Plotly boxplots for comparing a group of values with a reference standard
deviation value}
\usage{
boxplot_1sample_sigma(data, group, response, reflabel, reference, udm)
}
\arguments{
\item{data}{input data.frame with a column named *key* with progressive integers,
a column named *group* with a one-level factor label for the two groups
to be compared, a column named *response* with the numeric values for
the group and a column named *outlier* with a logical vector.}

\item{group}{a character string for the label of the grouping variable.}

\item{response}{a character string with the label for the response numeric variable.}

\item{reflabel}{a character value with the name of the reference value.}

\item{reference}{a numeric value with the reference mean.}

\item{udm}{a character string with the unit of measurement.}
}
\value{
A {plotly} boxplot for comparing a group of values with a reference
standard deviation value.
Raw data values are overlaid on top of the box.
}
\description{
The function provides a simple {plotly} boxplot for comparing
a group of values with a reference standard deviation value.
}
