% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compare031_2samples.R
\name{boxplot_2samples}
\alias{boxplot_2samples}
\title{Plotly boxplots for comparing two groups of values}
\usage{
boxplot_2samples(data, group, response, udm)
}
\arguments{
\item{data}{input data.frame with a column named *key* with progressive integers,
a column named *group* with a two level factor label for the two groups
to be compared, a column named *response* with the numeric values for
the two groups and a column named *outlier* with a logical vector.}

\item{group}{a character string for the label of the grouping variable.}

\item{response}{a character string with the label for the response numeric variable.}

\item{udm}{a character string with the unit of measurement.}
}
\value{
A {plotly} boxplot for comparing two group of values. Raw data values
are overlaid on top of the boxes.
}
\description{
The function provides a simple {plotly} boxplot for comparing
two groups of values
}
