% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compare032_2samples_par.R
\name{boxplot_2samples_par}
\alias{boxplot_2samples_par}
\title{Plotly boxplots for comparing two groups of values for one of which only mean,
standard deviation and number of samples are reported.}
\usage{
boxplot_2samples_par(data, group, response, udm, group2, dfsummary)
}
\arguments{
\item{data}{input data.frame with a column named *key* with progressive integers,
a column named *group* with a one level factor label, a column named *response*
with the numeric values and a column named *outlier* with a logical vector.}

\item{group}{a character string for the label of the grouping variable.}

\item{response}{a character string with the label for the response numeric variable.}

\item{udm}{a character string with the unit of measurement.}

\item{group2}{a character string for the name of the group for which only the
summary has been provided.}

\item{dfsummary}{a data.frame with mean, sd and number of values for the two
groups: it has three rows and three columns. Column names are "index",
name of the group1 and name of the group2.}
}
\value{
A {plotly} boxplot for comparing two group of values. Raw data values
are overlaid on top of the boxes.
}
\description{
The function provides a simple {plotly} boxplot for comparing
two groups of values for one of which only mean, standard deviation and number
of samples are reported.
}
