% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compare035_2values_unc.R
\name{boxplot_2values_unc}
\alias{boxplot_2values_unc}
\title{Plotly boxplots for comparing two values with extended uncertainties}
\usage{
boxplot_2values_unc(data, group, response, uncertainty, udm)
}
\arguments{
\item{data}{input data.frame with a column named *key* with progressive integers,
a column named *group* with a one-level factor label for the two groups
to be compared, a column named *response* with the numeric values for
the group and a column named *outlier* with a logical vector.}

\item{group}{a character string for the label of the grouping variable.}

\item{response}{a character string with the label for the response numeric variable.}

\item{uncertainty}{a character string with the label for the uncertainty numeric variable.}

\item{udm}{a character string with the unit of measurement.}
}
\value{
A {plotly} boxplot for comparing a group of values with a reference value.
Raw data values are overlaid on top of the box.
}
\description{
The function provides a simple {plotly} boxplot for comparing
two values with extended uncertainties.
}
