% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_estimate033_recuno.R
\name{boxplot_recuno}
\alias{boxplot_recuno}
\title{Plotly boxplots for comparing two values with extended uncertainties}
\usage{
boxplot_recuno(data, response, uncertainty, refvalue, refuncertainty, udm)
}
\arguments{
\item{data}{input data.frame with a column named *response* with the numeric values
and a column named *outlier* with a logical vector.}

\item{response}{a character string with the label for the response numeric variable.}

\item{uncertainty}{a character string with the label for the uncertainty numeric variable.}

\item{refvalue}{a \code{numeric} reference value.}

\item{refuncertainty}{the \code{numeric} uncertainty of the reference value.}

\item{udm}{a character string with the unit of measurement.}
}
\value{
A {plotly} boxplot for comparing a group of values with a reference value.
Raw data values are overlaid on top of the box.
}
\description{
The function provides a simple {plotly} boxplot for comparing
two values with extended uncertainties.
}
