% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_estimate031_riprec.R
\name{boxplot_riprec}
\alias{boxplot_riprec}
\title{Plotly boxplots for a a serie of measurement values.}
\usage{
boxplot_riprec(data, response, udm)
}
\arguments{
\item{data}{input data.frame with a column named *key* with progressive integers,
a column named *response* with the numeric values for the two groups and a
column named *outlier* with a logical vector.}

\item{response}{a character string with the label for the response numeric variable.}

\item{udm}{a character string with the unit of measurement.}
}
\value{
A {plotly} boxplot for measurement values.
}
\description{
The function provides a simple {plotly} boxplot for a serie
 of measurement values.
}
