% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_estimate031_riprec.R
\name{confint_riprec}
\alias{confint_riprec}
\title{Plotly confidence interval plot for a a serie of data and
a reference value with extended uncertainty.}
\usage{
confint_riprec(data, response, refvalue, refuncertainty, conflevel, udm)
}
\arguments{
\item{data}{input data.frame with a column named *key* with progressive integers,
a column named *response* with the numeric values for the two groups and a
column named *outlier* with a logical vector.}

\item{response}{a character string with the label for the response numeric variable.}

\item{refuncertainty}{the extended uncertainty for the reference numeric value.}

\item{conflevel}{confidence level for the mean confidence interval: allowed values
are 0.90, 0.95 and 0.99.}

\item{udm}{a character string with the unit of measurement.}

\item{refval}{a reference numeric value.}
}
\value{
A {plotly} interval plot for comparing the confidence
interval of the supplied  measurement values with a reference value and
its extended uncertainty.
}
\description{
The function provides a simple {plotly} for comparing the
confidence interval with a reference value and its extended uncertainty.
}
