% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_loadfile02.R
\name{csvimport}
\alias{csvimport}
\title{Importing CSV files}
\usage{
csvimport(mypath)
}
\arguments{
\item{mypath}{a string with the path of the csv file to be imported.}
}
\value{
a {data.table} with the imported data. String are imported as factors.
}
\description{
The function imports csv files trying to identify the
field and decimal separator. Intended csv files supported can have
 \itemize{
   \item{a comma as field separator and a dot as decimal separator;}
   \item{a semicolon as field separator and a comma as decimal separator;}
   \item{a semicolon as field separator and a dot as decimal separator.}
 }
 All files to be imported must have strings as column headings.
}
