% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compare034_1sample_sigma.R
\name{fct_chitest_1sample_sigma}
\alias{fct_chitest_1sample_sigma}
\title{Displays the results of a \eqn{\chi^2}-test for a group of values vs a
reference standard deviation value.}
\usage{
fct_chitest_1sample_sigma(
  data,
  response,
  group,
  reflabel,
  reference,
  significance = 0.95,
  alternative = "different"
)
}
\arguments{
\item{data}{a \code{data.frame} or \code{data.table} with the results
relevant for testing. A single-levels grouping \code{factor} variable
and a \code{numeric} vector with the measurements should be included.}

\item{response}{the name of a numeric vector in \code{data}.
Quotation (" ") is not required.}

\item{group}{the name of a single-level factor variable that identifies
the group in \code{data}. Quotation (" ") is not required.}

\item{reflabel}{a character value with the name of the reference value.}

\item{reference}{a numeric value with the reference standard deviation.}

\item{significance}{a number, typically either 0.90, 0.95 (default) or 0.99
indicating the confidence level for the test.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of \code{"different"} or \code{"greater"}.}
}
\value{
A list with the following items:
 \describe{
   \item{hypotheses}{a named vector of strings, being \code{h0} and \code{h1}
   the null and alternative hypothesis, respectively.}
   \item{ratio}{a named vector of numbers, being \code{sd},
   \code{lwrci} and \code{uprci} the variance of the provided numeric values and
   the lower and upper ends of the confidence interval, respectively.
   The confidence interval is calculated considering both the \code{significance}
   and \code{alternative} arguments. For \code{alternative = "greater"} only the
   lower end of the confidence interval will be calculated.}
   \item{test}{a named vector of numbers, being \code{dof}, \code{chisper},
   \code{chitheo} and \code{pvalue} the degrees of freedom, the calculated value
   of the \eqn{\chi^2} statistic, the tabulated value of the \eqn{\chi^2} statistic and
   the \eqn{p}-value associated to the test.}
   \item{result}{a string indicating whether H0 should be rejected or not.}
 }
}
\description{
The function displays the results of a \eqn{\chi^2}-test performed
 on a group of values vs a reference standard deviation value.
 The returned text is suitable for the {SI confronta} {shiny} app.
}
\details{
\eqn{\chi^2}-test is with the numerically larger variance as numerator.
 As a consequence, the alternative  hypothesis tested can only be
 \eqn{\mathrm{Var}(A) \neq \mathrm{Var}(B)}  (\code{alternative = "different"})
 or \eqn{\mathrm{Var}(A) > \mathrm{Var}(B)} (\code{alternative = "greater"}).
}
