% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_estimate031_riprec.R
\name{fct_shapiro}
\alias{fct_shapiro}
\title{Displays the results of a Shapiro-Wilk normality test}
\usage{
fct_shapiro(values)
}
\arguments{
\item{values}{a \code{vector} with the values relevant for testing.}
}
\value{
A list with the following items:
 \describe{
   \item{W}{a numeric value with the test statistic.}
   \item{pvalue}{a numeric value with the p-value of the test.}
   \item{result}{A string with the result of the test.}
 }
}
\description{
The function displays the results of a Shapiro-Wilk test for
 normality.
 The returned text is suitable for the {SI precisa} {shiny} app.
}
\details{
the Shapiro-Wilk test is calculated using the base-R function
 \code{shapiro.test}.
}
