% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_estimate033_recuno.R
\name{fct_trueness_recuno}
\alias{fct_trueness_recuno}
\title{Displays the results for the determination of trueness performances for a
single measurement value.}
\usage{
fct_trueness_recuno(data, response, refvalue)
}
\arguments{
\item{data}{a \code{data.frame} or \code{data.table} with the results
relevant for testing. At least a \code{numeric} vector with the
measurements should be included.}

\item{response}{the name of a numeric vector in \code{data}.}

\item{refvalue}{a reference value for the set of measurement values.}
}
\value{
A list with the following items:
 \describe{
   \item{recovery}{a numeric value with the relative recovery for the
                   mean measurement values.}
   \item{bias}{a numeric value with the bias for the mean measurement values.}
   \item{bias_rel}{a numeric value with the relative bias for the mean measurement values.}
 }
}
\description{
The function displays the results for the determination
of trueness performances of an analytical method, based on a single
measurement value.
The returned text is suitable for the {SI precisa} {shiny} app.
}
\details{
trueness expressed as recovery and bias are calculated according to
the Eurachem guide - The Fitness for Purpose of Analytical Methods.
}
