% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compare031_2samples.R
\name{fct_ttest_2samples}
\alias{fct_ttest_2samples}
\title{Displays the results of a \eqn{t}-test for two groups of values}
\usage{
fct_ttest_2samples(
  data,
  response,
  group,
  significance = 0.95,
  alternative = "different"
)
}
\arguments{
\item{data}{a \code{data.frame} or \code{data.table} with the results
relevant for testing. At least a two-levels grouping \code{factor} variable
and a \code{numeric} vector with the measurements should be included.}

\item{response}{the name of a numeric vector in \code{data}.
Quotation (" ") is not required.}

\item{group}{the name of a two-level factor variable that identifies the groups
in \code{data}. Quotation (" ") is not required.}

\item{significance}{a number, typically either 0.90, 0.95 (default) or 0.99
indicating the confidence level for the test.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of \code{"different"} or \code{"greater"}.}
}
\value{
A list with the following items:
 \describe{
   \item{hypotheses}{a named vector of strings, being \code{h0} and \code{h1}
   the null and alternative hypothesis, respectively.}
   \item{difference}{a named vector of numbers, being \code{mean},
   \code{lwrci} and \code{uprci} the difference in means of the two groups and
   the lower and upper ends of the confidence interval, respectively.
   The confidence interval is calculated considering both the \code{significance}
   and \code{alternative} arguments. For \code{alternative = "greater"} only the
   lower end of the confidence interval will be calculated.}
   \item{test}{a named vector of numbers, being \code{dof}, \code{tsper},
   \code{ttheo} and \code{pvalue} the degrees of freedom, the calculated value
   of the \eqn{t} statistic, the tabulated value of the \eqn{t} statistic and
   the \eqn{p}-value associated to the test. As in the original \code{t.test}
   function in base R, the statistic is calculated by performing a Welch test
   and approximating the actual number of degrees of freedom.}
   \item{result}{a string indicating whether H0 should be rejected or not.}
 }
}
\description{
The function displays the results of a \eqn{t}-test performed
 on two groups of values.
 The returned text is suitable for the {SI confronta} {shiny} app.
}
\details{
\eqn{t}-test is calculated using the base-R function \code{t.test},
 with the group with the numerically larger mean as first argument.
 As a consequence, for the means of two series of values \eqn{A} and \eqn{B},
 being the first one numerically greater than the second one, the alternative
 hypothesis tested can only be \eqn{\bar{A} \neq \bar{B}}
 (\code{alternative = "different"}) or \eqn{\bar{A} > \bar{B}}
 (\code{alternative = "greater"}).
}
