% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_estimate031_riprec.R
\name{fct_ttest_riprec}
\alias{fct_ttest_riprec}
\title{Displays the results of a \eqn{t}-test for a group of values vs a reference value.}
\usage{
fct_ttest_riprec(data, response, refvalue, significance = 0.95)
}
\arguments{
\item{data}{a \code{data.frame} or \code{data.table} with the results
relevant for testing. At least a \code{numeric} vector with the
measurements should be included.}

\item{response}{the name of a numeric vector in \code{data}.
Quotation (" ") is not required.}

\item{refvalue}{reference value for the comparison.}

\item{significance}{a number, typically either 0.90, 0.95 (default) or 0.99
indicating the confidence level for the test.}
}
\value{
A list with the following items:
 \describe{
   \item{hypotheses}{a named vector of strings, being \code{h0} and \code{h1}
   the null and alternative hypothesis, respectively.}
   \item{mean}{a named vector of numbers, being \code{mean},
   \code{lwrci} and \code{uprci} the mean of the measurement value, and
   the lower and upper end of its confidence interval, respectively.}
   \item{test}{a named vector of numbers, being \code{dof}, \code{tsper},
   \code{ttheo} and \code{pvalue} the degrees of freedom, the calculated value
   of the \eqn{t} statistic, the tabulated value of the \eqn{t} statistic and
   the \eqn{p}-value associated to the test. As in the original \code{t.test}
   function in base R, the statistic is calculated by performing a Student t-test}
   \item{result}{a string indicating whether H0 should be rejected or not.}
 }
}
\description{
The function displays the results of a \eqn{t}-test performed
 on a group of values vs a reference value.
 The returned text is suitable for the {SI precisa} {shiny} app.
}
\details{
\eqn{t}-test is calculated using the base-R function \code{t.test}
 with a two-sided alternative hypothesis.
}
