% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ftest_reference}
\alias{ftest_reference}
\title{F-Test data}
\format{
a dataframe with 5 rows and 2 columns:
\describe{
   \item{group}{the grouping variable, either \code{a} or \code{b}}
   \item{value}{the measured value for the variable of interest}
}
}
\source{
\url{https://support.microsoft.com/en-us/office/f-test-function-100a59e7-4108-46f8-8443-78ffacb6c0a7}.
}
\usage{
ftest_reference
}
\description{
A dataset with 10 values divided in two groups of 5 values. The dataset is
provided for testing the \code{fct_test} results.
}
\keyword{data}
