% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compare034_1sample_sigma.R
\name{ggboxplot_1sample_sigma}
\alias{ggboxplot_1sample_sigma}
\title{GGplot2 boxplots for comparing a group of values with a standard deviation
reference value}
\usage{
ggboxplot_1sample_sigma(data, group, response, reflabel, reference, udm)
}
\arguments{
\item{data}{input data.frame with a column named *key* with progressive integers,
a column with a two level factor label for the two groups
to be compared, a column with the numeric values for the two groups and a
column named *rimosso* with "sì" or "no" values.}

\item{group}{a character string for the label of the grouping variable.}

\item{response}{a character string with the label for the response numeric variable.}

\item{reflabel}{a character value with the name of the reference value.}

\item{reference}{a numeric value with the reference mean.}

\item{udm}{a character string with the unit of measurement.}
}
\value{
A {ggplot2} boxplot for comparing a group of values with a reference value.
Raw data values are overlaid on top of the box.
}
\description{
The function provides a simple {ggplot2} boxplot for comparing
a group of values with a standard deviation reference value
}
