% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compare032_2samples_par.R
\name{ggboxplot_2samples_par}
\alias{ggboxplot_2samples_par}
\title{GGplot2 boxplots for comparing two groups of values, for one of which only mean,
standard deviation and number of samples are reported.}
\usage{
ggboxplot_2samples_par(data, group, response, udm, group2, dfsummary)
}
\arguments{
\item{data}{input data.frame with a column named *key* with progressive integers,
a column with a two level factor label for the two groups
to be compared, a column with the numeric values for the two groups and a
column named *rimosso* with "sì" or "no" values.}

\item{group}{a character string for the label of the grouping variable.}

\item{response}{a character string with the label for the response numeric variable.}

\item{udm}{a character string with the unit of measurement.}

\item{group2}{a character string for the name of the group for which only the
summary has been provided.}

\item{dfsummary}{a data.frame with mean, sd and number of values for the two
groups: it has three rows and three columns. Column names are "index",
name of the group1 and name of the group2.}
}
\value{
A {ggplot2} boxplot for comparing two group of values. Raw data values
are overlaid on top of the boxes.
}
\description{
The function provides a simple {ggplot2} boxplot for comparing
two groups of values for one of which only mean, standard deviation and number
of samples are reported.
}
