% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compare035_2values_unc.R
\name{ggboxplot_2values_unc}
\alias{ggboxplot_2values_unc}
\title{GGplot2 boxplots for comparing two values with extended uncertainties}
\usage{
ggboxplot_2values_unc(data, group, response, uncertainty, udm)
}
\arguments{
\item{data}{input data.frame with a column named *key* with progressive integers,
a column with a two level factor label for the two groups
to be compared, a column with the numeric values for the two groups and a
column named *rimosso* with "sì" or "no" values.}

\item{group}{a character string for the label of the grouping variable.}

\item{response}{a character string with the label for the response numeric variable.}

\item{uncertainty}{a character string with the label for the uncertainty numeric variable.}

\item{udm}{a character string with the unit of measurement.}
}
\value{
A {ggplot2} boxplot for comparing two values with extended uncertainties.
Raw data values are overlaid on top of the box.
}
\description{
The function provides a simple {ggplot2} boxplot for comparing
two values with extended uncertainties.
}
