% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_estimate033_recuno.R
\name{ggboxplot_recuno}
\alias{ggboxplot_recuno}
\title{GGplot2 boxplots for comparing two values with extended uncertainties}
\usage{
ggboxplot_recuno(data, response, uncertainty, refvalue, refuncertainty, udm)
}
\arguments{
\item{data}{input data.frame with a column named *key* with progressive integers,
a column with the numeric values for the two groups and a column named *rimosso*
with "sì" or "no" values.}

\item{response}{a character string with the label for the response numeric variable.}

\item{uncertainty}{a character string with the label for the uncertainty numeric variable.}

\item{refvalue}{a \code{numeric} reference value.}

\item{refuncertainty}{the \code{numeric} uncertainty of the reference value.}

\item{udm}{a character string with the unit of measurement.}
}
\value{
A {ggplot2} boxplot for comparing two values with extended uncertainties.
Raw data values are overlaid on top of the box.
}
\description{
The function provides a simple {ggplot2} boxplot for comparing
two values with extended uncertainties.
}
