% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_estimate031_riprec.R
\name{ggboxplot_riprec}
\alias{ggboxplot_riprec}
\title{ggplot2 boxplots for a a serie of data, and for comparing the confidence interval
with a reference value and its extended uncertainty}
\usage{
ggboxplot_riprec(data, response, refvalue, refuncertainty, udm)
}
\arguments{
\item{data}{input data.frame a numeric columns with the response and a
character columns with outliers flag.
The first column must be named as the response variable, while the second column
must be named *rimosso*.}

\item{response}{a character string with the label for the response numeric variable.}

\item{refuncertainty}{the extended uncertainty for the reference numeric value.}

\item{udm}{a character string with the unit of measurement.}

\item{refval}{a reference numeric value.}
}
\value{
A {ggplot2} boxplot and an interval plot for comparing the confidence
interval of the supplied  measurement values with a reference value and
its extended uncertainty.
}
\description{
The function provides a simple {ggplot2} boxplot for a serie
 of data, and for comparing the confidence interval with a reference value
 and its extended uncertainty.
}
