% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compare033_1sample_mu.R
\name{rowsummary_1sample_mu}
\alias{rowsummary_1sample_mu}
\title{Summary arranged on rows for one group vs a mean reference value}
\usage{
rowsummary_1sample_mu(
  data,
  response,
  group,
  reflabel,
  reference,
  udm = "",
  signif = 3L
)
}
\arguments{
\item{data}{the \code{data.frame} or \code{data.table} to be summarised.}

\item{response}{a string with the name of the variable to summarise.}

\item{group}{a string with the name of the grouping variable.}

\item{reflabel}{a character value with the name of the reference value.}

\item{reference}{a numeric value with the reference mean.}

\item{udm}{a string with the unit of measurement.}

\item{signif}{an integer with the number of desired significant figures.}
}
\value{
a \code{data.table} with 6 rows and \eqn{n + 1} columns for \eqn{n}
  levels of the group variable.
}
\description{
The function returns a table with max, mean, median, min, sd and n
 values arranged on rows whereas groups are on columns. Numbers are formatted as
 text in order to provide the desired significant figures.
}
