% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compare032_2samples_par.R
\name{rowsummary_2samples_par}
\alias{rowsummary_2samples_par}
\title{Summary arranged on rows for two groups for one of which only mean, standard deviation
and number of samples are reported.}
\usage{
rowsummary_2samples_par(
  data,
  response,
  group,
  group2,
  mean2,
  sd2,
  n2,
  udm = "",
  signif = 3L
)
}
\arguments{
\item{data}{the \code{data.frame} or \code{data.table} to be summarised.}

\item{response}{a string with the name of the variable to summarise.}

\item{group}{a string with the name of the grouping variable.}

\item{group2}{a character value with the name of the second group.}

\item{mean2}{a numeric value with the mean for the second group.}

\item{sd2}{a numeric value with the standard deviation for the second group.}

\item{n2}{a numeric value with the number of values for the second group.}

\item{udm}{a string with the unit of measurement.}

\item{signif}{an integer with the number of desired significant figures.}
}
\value{
a \code{data.table} with 6 rows and \eqn{n + 1} columns for \eqn{n}
  levels of the group variable.
}
\description{
The function returns a table with max, mean, median, min, sd and n
 values arranged on rows whereas groups are on columns. Numbers are formatted as
 text in order to provide the desired significant figures.
}
