% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_compare035_2values_unc.R
\name{rowsummary_2values_unc}
\alias{rowsummary_2values_unc}
\title{Summary arranged on rows for two values with their extended uncertainties}
\usage{
rowsummary_2values_unc(
  data,
  response,
  uncertainty,
  group,
  udm = "",
  signif = 3L
)
}
\arguments{
\item{data}{the \code{data.frame} or \code{data.table} to be summarised.}

\item{response}{a string with the name of the variable with the measurement values.}

\item{uncertainty}{a string with the name of the variable with the uncertainty values.}

\item{group}{a string with the name of the grouping variable.}

\item{udm}{a string with the unit of measurement.}

\item{signif}{an integer with the number of desired significant figures.}
}
\value{
a \code{data.table} with 2 rows and 2 columns.
}
\description{
The function returns a table with measurement values and their
 extended uncertainties. Numbers are formatted as text in order to provide
 the desired significant figures.
}
