% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_estimate033_recuno.R
\name{rowsummary_recuno}
\alias{rowsummary_recuno}
\title{Summary arranged on rows for two values with their extended uncertainties}
\usage{
rowsummary_recuno(
  data,
  response,
  uncertainty,
  refvalue,
  refuncertainty,
  udm = "",
  signif = 3L
)
}
\arguments{
\item{data}{the \code{data.frame} or \code{data.table} to be summarised.}

\item{response}{a string with the name of the variable with the measurement values.}

\item{uncertainty}{a string with the name of the variable with the uncertainty values.}

\item{refvalue}{a \code{numeric} reference value.}

\item{refuncertainty}{the \code{numeric} uncertainty of the reference value.}

\item{udm}{a string with the unit of measurement.}

\item{signif}{an integer with the number of desired significant figures.}
}
\value{
a \code{data.frame} with 2 rows and 2 columns.
}
\description{
The function returns a table with measurement and reference values,
 and their extended uncertainties.
 Numbers are formatted as text in order to provide the desired significant figures.
}
