% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_estimate032_rip.R
\name{rowsummary_rip}
\alias{rowsummary_rip}
\title{Summary arranged on rows for a set of measurements and a reference value}
\usage{
rowsummary_rip(data, response, udm = "", signif = 3L)
}
\arguments{
\item{data}{the \code{data.frame} or \code{data.table} to be summarised a
column named *outlier* with a logical flag must be included.}

\item{response}{a character with the name of the data columns with the measured values.}

\item{udm}{a string with the unit of measurement.}

\item{signif}{an integer with the number of desired significant figures.}
}
\value{
a \code{data.table} with 6 rows and \eqn{n + 1} columns for \eqn{n}
  levels of the group variable.
}
\description{
The function returns a table with max, mean, median, min, n and
 the number of removed points.
 values are arranged on rows, columns are for measurement and reference values,
 respectively . Numbers are formatted as text in order to provide the
 desired significant figures.
}
