% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shapiro_reference}
\alias{shapiro_reference}
\alias{shapirotest_reference}
\title{Shapiro-Wilk test data}
\format{
a vector with 11 numerical elements.
}
\source{
An analysis of variance test for normality (complete samples),
 Biometrika (1965), 52, 3 and 2, p. 591.
 Section 4 - Examples, pag. 606, Example 1.
 \url{http://links.jstor.org/sici?sici=0006-3444\%28196512\%2952\%3A3\%2F4\%3C591\%3AAAOVTF\%3E2.0.CO\%3B2-B}.
}
\usage{
shapirotest_reference
}
\description{
A dataset with 11 values of men weights expressed in pounds.
The dataset is provided for testing the results of \code{fct_shapiro}.
}
\author{
S. S. Shapiro

M. B. Wilk
}
\keyword{data}
