% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_estimate032_rip.R
\name{shewart_rip}
\alias{shewart_rip}
\title{Plotly widget for differences between pair of values vs their median values.}
\usage{
shewart_rip(data, measure1, measure2, udm = "")
}
\arguments{
\item{data}{input data.frame with a column named *key* with progressive integers,
two columns named *measure1* and *measure2* with the numeric values
of the paired measurement values, and a column named *outlier* with a logical vector.}

\item{measure1}{a character string with the label for the first response
numeric variable.}

\item{measure2}{a character string with the label for the second response
numeric variable.}

\item{udm}{a character string with unit of measurement of the paired values.
Default is an empty string.}
}
\value{
A {plotly} interval plot for comparing the confidence
interval of the supplied  measurement values with a reference value and
its extended uncertainty.
}
\description{
The function provides a simple {plotly} for
differences between pair of values vs their mean values.
}
\details{
relative percentage difference between values are plotted vs the mean of the
paired values. The central line is calculated as the mean of the relative
differences and the upper limit is the calculated by multiplying the value of
the central line by 3.267, as described in section 6.4 of UNI ISO 7870-2:2023.
}
