% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tomato_yields}
\alias{tomato_yields}
\title{Tomato yields for two fertilizer mixtures}
\format{
a dataframe with 6 rows and 3 columns:
\describe{
   \item{parameter}{the parameter observed in the experiment, the yield}
   \item{fertilizer}{the variable for the useed fertilizer mixture: \code{a} or \code{b}}
   \item{pounds}{the measured yields, in pounds}
}
}
\source{
Statistics for Experimenters, Design, Innovation and Discovery.
 Wiley, Second Edition, 2005. ISBN: 978-0-471-71813-0.
 Section 3.1, pag. 78, table 3.3
}
\usage{
tomato_yields
}
\description{
A dataset containing the results expressed in pounds for tomato yields obtained
 by a randomized experiment with two fertilizer mixtures. The variable are as
 follows:
}
\author{
George E. P. Box

J. Stuart Hunter

William G. Hunter
}
\keyword{data}
