% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniiso_11352_b1}
\alias{uniiso_11352_b1}
\title{Dissolved orthophosphate concentration values from UNI ISO 11352:2012, Table B.1.}
\format{
a dataframe with 30 rows and 2 columns:
\describe{
   \item{id}{a numeric progressive identifier.}
   \item{value}{the measured orthophosphate concentration values, expressed as micromoles per liter.}
}
}
\source{
UNI ISO 11352:2012, Table B.1, pag. 17.
\url{https://store.uni.com/iso-11352-2012}
}
\usage{
uniiso_11352_b1
}
\description{
A dataset with 30 numerical values. The dataset is
provided for testing the \code{fct_precision} results.
}
\keyword{data}
