% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniiso_11352_b4}
\alias{uniiso_11352_b4}
\title{UNI ISO 11352:2012 Table B.4}
\format{
a dataframe with 10 rows and 1 column of numeric values named
{rel_diff}.
}
\source{
\url{https://store.uni.com/iso-11352-2012}
}
\usage{
uniiso_11352_b4
}
\description{
A dataset with 10 values of relative differences from paired values.
The dataset is provided for testing the \code{fct_precision_rip} results.
}
\keyword{data}
