% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniiso_16269_4_432}
\alias{uniiso_16269_4_432}
\title{Observations from UNI ISO 16269-4:2019 - Section 4.3.2}
\format{
a vector with 20 numerical values:
}
\source{
UNI ISO 16269-4:2019 - Statistical interpretation of data - Part 4:
 Detection and treatment of outliers. Section 4.3.2.
 \url{https://store.uni.com/uni-iso-16269-4-2019}.
}
\usage{
uniiso_16269_4_432
}
\description{
A dataset containing 20 observations for outliers detection.
The variable are as follows:
}
\author{
ISO/TC 69 - Applications of statistical methods
}
\keyword{data}
