% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniiso_2854_x}
\alias{uniiso_2854_x}
\title{UNI ISO 2854:1988 Section X}
\format{
a dataframe with 12 rows and 2 columns:
\describe{
   \item{group}{the grouping variable, either \code{a} or \code{b}}
   \item{value}{the breaking load numerical values for the tested yarns, expressed in Newton.}
}
}
\source{
\url{https://store.uni.com/uni-iso-2854-1988}
}
\usage{
uniiso_2854_x
}
\description{
A dataset with 22 values divided in two groups with 10 and 12 values, respectively.
The dataset is provided for testing the \code{fct_test} results.
}
\keyword{data}
