% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniiso_5725_4_b3}
\alias{uniiso_5725_4_b3}
\title{Manganese mean concentration values from UNI ISO 5725-4:2020, Table B.3.}
\format{
a dataframe with 12 rows and 2 columns:
\describe{
   \item{lab}{a numeric value for the mean value for each laboratory measurements.}
   \item{value}{the percentage mean of the measured manganese contents provided by each laboratory}
}

a dataframe with 12 rows and 2 columns:
\describe{
   \item{lab}{a numeric value for the mean value for each laboratory measurements.}
   \item{value}{the percentage mean of the measured manganese contents provided by each laboratory}
}
}
\source{
UNI ISO 5725-4:2020, Table B.3, Content level 3, pag. 20.
\url{https://store.uni.com/uni-iso-5725-4-2020}

UNI ISO 5725-4:2020, Table B.3, Content level 3, pag. 20.
\url{https://store.uni.com/uni-iso-5725-4-2020}
}
\usage{
uniiso_5725_4_b3

uniiso_5725_4_b3
}
\description{
A dataset with 12 numerical values. The dataset is
provided for testing the \code{fct_precision} and \code{fct_trueness} results.

A dataset with 12 numerical values. The dataset is
provided for testing the \code{fct_precision} and \code{fct_trueness} results.
}
\keyword{data}
