% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uranium_cps}
\alias{uranium_cps}
\title{Mass spectrometer measurements on a Uranium isotope}
\format{
a vector with 8 numerical values:
}
\source{
Some Grubbs-Type Statistics for the Detection of Several Outliers.
 Technometrics, 14(3), 1972, pp. 583-597.
 Also available at
 \url{https://www.itl.nist.gov/div898/handbook/eda/section3/eda35h1.htm}.
}
\usage{
uranium_cps
}
\description{
A dataset containing the results expressed counts per seconds for mass spectrometer measurements on a Uranium isotope.
The variable are as follows:
}
\author{
Gary L. Tietjen

Roger H. Moore
}
\keyword{data}
