% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_estimate031_riprec.R
\name{boxplot_riprec}
\alias{boxplot_riprec}
\title{Plotly boxplots for a a serie of data, and for comparing the confidence interval
with a reference value and its extended uncertainty}
\usage{
boxplot_riprec(data, response, refvalue, refuncertainty, udm)
}
\arguments{
\item{data}{input data.frame with a column named *key* with progressive integers,
a column named *response* with the numeric values for the two groups and a
column named *outlier* with a logical vector.}

\item{response}{a character string with the label for the response numeric variable.}

\item{refuncertainty}{the extended uncertainty for the reference numeric value.}

\item{udm}{a character string with the unit of measurement.}

\item{refval}{a reference numeric value.}
}
\value{
A {plotly} boxplot and an interval plot for comparing the confidence
interval of the supplied  measurement values with a reference value and
its extended uncertainty.
}
\description{
The function provides a simple {plotly} boxplot for a serie
 of data, and for comparing the confidence interval with a reference value
 and its extended uncertainty.
}
