% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_estimate032_rip.R
\name{ggboxplot_rip}
\alias{ggboxplot_rip}
\title{ggplot2 boxplots for paired measurement values.}
\usage{
ggboxplot_rip(data, response, second_response)
}
\arguments{
\item{data}{input data.frame a numeric columns with the two numeric responses and a
character columns with outliers flag.
The first two columns must be named as the response and second_response variable,
respectively, while the second column must be named *rimosso*.}

\item{response}{a character string with the label for the first response numeric variable.}

\item{second_response}{a character string with the label for the second response numeric variable.}

\item{udm}{a character string with the unit of measurement.}
}
\value{
A {ggplot2} boxplot for the differences in paired measurement values.
}
\description{
The function provides a simple {ggplot2} boxplot for a serie
 of data, and for comparing the confidence interval with a reference value
 and its extended uncertainty.
}
